Introducing SubFilterSharedViewModel to host Event bus for subfilters and mainly bottom sheet visibility for all the reader tabs. SubFilterViewModel will be dedicated per each tab. Organization is used to discriminate which tab we are in if Filterable.
